# docs/openapi/community-map-api.openapi.yaml
openapi: 3.1.0
info:
  title: Chengdu Tongzilin Community Map API
  version: 0.1.0
  description: |
    Community Map API for mobile, Mini Program, Admin, and Postman debugging.

    Current status:
    - Local Koa HTTP routes and shared contracts are implemented for the listed endpoints.
    - CloudBase dev URL is documented for deployment target only; `/api` route and live acceptance still require verification.
    - `x-mock-user-id` is a local/dev actor header and is not a production authentication scheme.
servers:
  - url: http://127.0.0.1:8787
    description: local Koa dev server
  - url: https://cloud1-d7gxdk8t43bd639c0.service.tcloudbase.com/api
    description: CloudBase dev HTTP access service, pending `/api` route verification
...
components:
  securitySchemes:
    MockActorHeader:
      type: apiKey
      in: header
      name: x-mock-user-id
      description: Local/dev mock actor id, for example `user_001` for admin flows.
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 10
    CommunityId:
      name: communityId
      in: query
      schema:
        type: string
        default: tongzilin
    Keyword:
      name: keyword
      in: query
      schema:
        type: string
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples:
            [
              VALIDATION_ERROR,
              UNAUTHORIZED,
              FORBIDDEN,
              NOT_FOUND,
              INTERNAL_ERROR
            ]
        message:
          type: string
        details: {}
    ApiFailure:
      type: object
      required: [success, error, requestId]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
    SuccessEnvelope:
      type: object
      required: [success, data, requestId]
      properties:
        success:
          type: boolean
          const: true
        data: {}
        requestId:
          type: string
...
    Post:
      type: object
      required:
        [
          _id,
          author_user_id,
          author_display,
          community_id,
          title,
          content,
          language,
          tag_ids,
          location_text,
          image_file_ids,
          image_urls,
          place_id,
          event_id,
          comment_count,
          like_count,
          favorite_count,
          share_count,
          created_at,
          updated_at,
          status,
          review_status
        ]
      properties:
        _id:
          type: string
        author_user_id:
          type: string
        author_display:
          type: object
          required: [nickname, avatar_url]
          properties:
            nickname:
              type: string
            avatar_url:
              type:
                - string
                - "null"
              format: uri
        community_id:
          type: string
        title:
          type: string
        content:
          type: string
        language:
          type: string
          enum: [zh, en]
        tag_ids:
          type: array
          items:
            type: string
        location_text:
          type:
            - string
            - "null"
        image_file_ids:
          type: array
          items:
            type: string
        image_urls:
          type: array
          items:
            type: string
            format: uri
        place_id:
          type:
            - string
            - "null"
        event_id:
          type:
            - string
            - "null"
        comment_count:
          type: integer
          minimum: 0
        like_count:
          type: integer
          minimum: 0
        favorite_count:
          type: integer
          minimum: 0
        share_count:
          type: integer
          minimum: 0
        created_at:
          type: string
        updated_at:
          type: string
        status:
          type: string
        review_status:
          type: string
    CreatePostInput:
      type: object
      required: [title, content, language, tag_ids]
      properties:
        title:
          type: string
        content:
          type: string
        language:
          type: string
          enum: [zh, en]
        tag_ids:
          type: array
          items:
            type: string
        location_text:
          type:
            - string
            - "null"
          default: null
        image_file_ids:
          type: array
          items:
            type: string
          default: []
        image_urls:
          type: array
          items:
            type: string
            format: uri
          default: []
        place_id:
          type:
            - string
            - "null"
          default: null
        event_id:
          type:
            - string
            - "null"
          default: null
    Comment:
      type: object
      required:
        [_id, post_id, author_user_id, content, language, status, created_at]
      properties:
        _id:
          type: string
        post_id:
          type: string
        author_user_id:
          type: string
        content:
          type: string
        language:
          type: string
          enum: [zh, en]
        status:
          type: string
        created_at:
          type: string
    CreateCommentInput:
      type: object
      required: [content, language]
      properties:
        content:
          type: string
        language:
          type: string
          enum: [zh, en]
    ReportPostInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
        description:
          type: string
    ModeratePostInput:
      type: object
      required: [review_status]
      properties:
        review_status:
          type: string
          enum: [visible, hidden, deleted]
        reason:
          type: string
    PermanentDeletePostResponse:
      type: object
      required: [post_id, audit_record_id, deleted]
      properties:
        post_id:
          type: string
        audit_record_id:
          type: string
        deleted:
          type: object
          additionalProperties: false
          properties:
            posts: { type: integer, minimum: 0 }
            comments: { type: integer, minimum: 0 }
            interactions: { type: integer, minimum: 0 }
            reports: { type: integer, minimum: 0 }
            notifications: { type: integer, minimum: 0 }
            file_assets: { type: integer, minimum: 0 }
            storage_objects: { type: integer, minimum: 0 }
            audit_records: { type: integer, minimum: 0 }
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
...
    Forbidden:
      description: Insufficient permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
paths:
...
  /discover/posts:
    get:
      tags: [Discover]
      operationId: listPosts
      summary: List discover posts
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/CommunityId"
        - $ref: "#/components/parameters/Keyword"
      responses:
        "200":
          description: Post page
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [items, page, pageSize, total]
                        properties:
                          items:
                            type: array
                            items:
                              $ref: "#/components/schemas/Post"
                          page:
                            type: integer
                          pageSize:
                            type: integer
                          total:
                            type: integer
    post:
      tags: [Discover]
      operationId: createPost
      summary: Create discover post
      security:
        - MockActorHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Post"
        "400":
          $ref: "#/components/responses/BadRequest"
  /discover/posts/{id}:
    get:
      tags: [Discover]
      operationId: getPost
      summary: Get post detail
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Post"
        "404":
          $ref: "#/components/responses/NotFound"
  /discover/me/posts:
    get:
      tags: [Discover]
      operationId: listMyPosts
      summary: List current actor discover posts
      security:
        - MockActorHeader: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/CommunityId"
      responses:
        "200":
          description: Current actor post page
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [items, page, pageSize, total]
                        properties:
                          items:
                            type: array
                            items:
                              $ref: "#/components/schemas/Post"
                          page:
                            type: integer
                          pageSize:
                            type: integer
                          total:
                            type: integer
  /discover/posts/{id}/comments:
    get:
      tags: [Discover]
      operationId: listPostComments
      summary: List visible post comments
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Comment page
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [items, page, pageSize, total]
                        properties:
                          items:
                            type: array
                            items:
                              $ref: "#/components/schemas/Comment"
                          page:
                            type: integer
                          pageSize:
                            type: integer
                          total:
                            type: integer
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Discover]
      operationId: createComment
      summary: Create post comment
      security:
        - MockActorHeader: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Comment"
  /discover/posts/{id}/report:
    post:
      tags: [Discover]
      operationId: reportPost
      summary: Report post
      security:
        - MockActorHeader: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportPostInput"
      responses:
        "200":
          description: Reported post
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Post"
  /admin/discover/posts/{id}:
    delete:
      tags: [Admin, Discover]
      operationId: permanentlyDeletePost
      summary: Permanently delete a post and its dependencies
      security:
        - MockActorHeader: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Permanent deletion result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PermanentDeletePostResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/discover/posts/{id}/moderation:
    post:
      tags: [Admin, Discover]
      operationId: adminModeratePost
      summary: Moderate post
      security:
        - MockActorHeader: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModeratePostInput"
      responses:
        "200":
          description: Moderated post
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Post"
        "403":
          $ref: "#/components/responses/Forbidden"
...